requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
networkx==3.2.1
sentence-transformers==2.2.2
chromadb==0.4.18
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
plotly==5.17.0
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2